+++ fdr-impl/src/main/resources/example.groovy/ExampleFilterProcessor.groovy
@@ -1,0 +1,34 @@
+import com.dwarfeng.dcti.stack.bean.dto.DataInfo
+import com.dwarfeng.fdr.impl.handler.filter.GroovyFilterMaker
+import com.dwarfeng.fdr.stack.bean.entity.FilteredValue
+import com.dwarfeng.fdr.stack.exception.FilterException
+import com.dwarfeng.subgrade.stack.bean.key.LongIdKey
+
+/**
+ * 通过DataInfo的值的长度判断数据信息是否通过过滤的脚本。
+ * <p> 如果DataInfo中数据的长度大于5，则不通过，否则通过。
+ */
+@SuppressWarnings("GrPackage")
+class ExampleFilterProcessor implements GroovyFilterMaker.Processor {
+
+    @Override
+    FilteredValue test(LongIdKey pointIdKey, LongIdKey filterIdKey, DataInfo dataInfo) throws FilterException {
+        try {
+            def size = dataInfo.getValue().size()
+            if (size > 5) {
+                return new FilteredValue(
+                        null,
+                        pointIdKey,
+                        filterIdKey,
+                        dataInfo.getHappenedDate(),
+                        dataInfo.getValue(),
+                        "DataInfo 的值大于 5 个字符"
+                )
+            } else {
+                return null
+            }
+        } catch (Exception e) {
+            throw new FilterException(e)
+        }
+    }
+}