+++ fdr-impl/src/main/resources/example.groovy/ExampleTriggerProcessor.groovy
@@ -1,0 +1,34 @@
+import com.dwarfeng.dcti.stack.bean.dto.DataInfo
+import com.dwarfeng.fdr.impl.handler.trigger.GroovyTriggerMaker
+import com.dwarfeng.fdr.stack.bean.entity.TriggeredValue
+import com.dwarfeng.fdr.stack.exception.TriggerException
+import com.dwarfeng.subgrade.stack.bean.key.LongIdKey
+
+/**
+ * 通过DataInfo的值的长度判断是否触发的脚本。
+ * <p> 如果DataInfo中数据的长度小于5，则触发，否则不触发。
+ */
+@SuppressWarnings("GrPackage")
+class ExampleTriggerProcessor implements GroovyTriggerMaker.Processor {
+
+    @Override
+    TriggeredValue test(LongIdKey pointIdKey, LongIdKey triggerIdKey, DataInfo dataInfo) throws TriggerException {
+        try {
+            def size = dataInfo.getValue().size()
+            if (size < 5) {
+                return new TriggeredValue(
+                        null,
+                        pointIdKey,
+                        triggerIdKey,
+                        dataInfo.getHappenedDate(),
+                        dataInfo.getValue(),
+                        "DataInfo 的值小于 5 个字符"
+                )
+            } else {
+                return null
+            }
+        } catch (Exception e) {
+            throw new TriggerException(e)
+        }
+    }
+}