+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 #### 功能构建
 
-- (无)
+- 实现Groovy过滤器以及Groovy触发器。
+  - com.dwarfeng.fdr.impl.handler.filter.GroovyFilterMaker
+  - com.dwarfeng.fdr.impl.handler.trigger.GroovyTriggerMaker
 
 #### Bug修复